pyproject.toml
[project]
name = "agentic-design-patterns"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.41.0",
    "python-dotenv>=1.1.1",
]
